core: Move state_producer to support
I don't want a dependency to core from interface, so need to do this.

Relates to #1138
