+++ docs/api/pyatv/interface.html
@@ -308,7 +308,7 @@ all its features.</p>
 <h3>Ancestors</h3>
 <ul class="hlist">
 <li>abc.ABC</li>
-<li>pyatv.core.StateProducer</li>
+<li>pyatv.support.state_producer.StateProducer</li>
 <li>typing.Generic</li>
 </ul>
 <h3>Subclasses</h3>
@@ -1142,7 +1142,7 @@ always be the same for the same content, but it is not guaranteed.</p></section>
 <h3>Ancestors</h3>
 <ul class="hlist">
 <li>abc.ABC</li>
-<li>pyatv.core.StateProducer</li>
+<li>pyatv.support.state_producer.StateProducer</li>
 <li>typing.Generic</li>
 </ul>
 <h3>Subclasses</h3>
@@ -1270,7 +1270,7 @@ always be the same for the same content, but it is not guaranteed.</p></section>
 <h3>Ancestors</h3>
 <ul class="hlist">
 <li>abc.ABC</li>
-<li>pyatv.core.StateProducer</li>
+<li>pyatv.support.state_producer.StateProducer</li>
 <li>typing.Generic</li>
 </ul>
 <h3>Subclasses</h3>
