+++ pyatv/protocols/mrp/connection.py
@@ -6,9 +6,9 @@ import logging
 
 from pyatv import exceptions
 from pyatv.core.net import tcp_keepalive
-from pyatv.interface import StateProducer
 from pyatv.protocols.mrp import protobuf
 from pyatv.support import chacha20, log_binary, log_protobuf
+from pyatv.support.state_producer import StateProducer
 from pyatv.support.variant import read_variant, write_variant
 
 _LOGGER = logging.getLogger(__name__)
